fixed various warnings in Stdlib build. The main one outstanding is FibreScanIO which compiles bison generated code

+++ src/stdio/src/FibreIO/FibrePrint.c
@@ -92,7 +92,7 @@ int FibreWriteAll(FILE *stream, int dim, int *shp, void *arr,
         } else {
           INDENT(stream, indent+3);
         }
-        fprintf(stream, "%f\n", ((float*)arr)[done + i]);
+        fprintf(stream, "%f\n", (double)((float*)arr)[done + i]);
         break;
       case DOUBLE:
         if((((double*)arr)[done + i]) >= 0) {
